Repository reorganization
The repository was reorganized to be more Arduino's library.

The code updated by @marcosdepo was moved to the src diretory, and the
repeated codes GSM_MQTT.h and GSM_MQTT.cpp from publish, and subscribe
directory were deleted because of obsolescence.

The keywords.txt was added to be more Arduino's library.

The README.md was updated asking for some good changes.